01-validated-basic-agent/pyproject.toml
[project]
name = "01-basic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "instructor>=1.6.3",
    "litellm>=1.52.3",
    "llm>=0.17.1",
    "openai>=1.51.2",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
